% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{md_.fa}
\alias{md_.fa}
\title{Markdown Lines for \link[psych]{fa} Object}
\usage{
\method{md_}{fa}(x, xnm, ...)
}
\arguments{
\item{x}{an \link[psych]{fa} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
Markdown Lines for \link[psych]{fa} Object
}
