% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{plot_fa_}
\alias{plot_fa_}
\title{Alternative Plot for \link[psych]{fa} Object}
\usage{
plot_fa_(x, ...)
}
\arguments{
\item{x}{an \link[psych]{fa} object}

\item{...}{..}
}
\description{
Alternative Plot for \link[psych]{fa} Object
}
\examples{
# ?psych::fa
m1 = Harman74.cor$cov |> 
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
m1a = Harman74.cor$cov |> 
 cov2cor() |>
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
# names(which(!mapply(adv.tzh::relaxed_identical, m1, m1a)))

m1 |> fa.plot() # ?psych::plot.psych; ?psych::fa.plot
m1 |> plot_fa_()

m2 = swiss |>
 cov() |> 
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
m2 |> fa.plot() 
m2 |> plot_fa_()
 
list(
 'fa1' = m1,
 'fa2' = m2
) |> fastmd::render2html()

}
\keyword{internal}
