% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa_ext}
\alias{fa_ext}
\alias{plot_fa_}
\alias{rmd_.fa}
\title{S3 Methods for \link[psych]{fa}}
\usage{
plot_fa_(x, ...)

rmd_.fa(x, xnm, ...)
}
\arguments{
\item{x}{an \link[psych]{fa} object}

\item{...}{..}

\item{xnm}{..}
}
\description{
S3 Methods for \link[psych]{fa}
}
\examples{
library(psych)
# ?psych::fa

m1 = Harman74.cor$cov |> 
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
m1a = Harman74.cor$cov |> 
 cov2cor() |>
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
# names(which(!mapply(ThomasJeffersonUniv::relaxed_identical, m1, m1a)))

m1 |> fa.plot() # ?psych::plot.psych; ?psych::fa.plot

m2 = swiss |>
 cov() |> 
 fa(nfactors = 2L, fm = 'pa', rotate = 'varimax')
m2 |> fa.plot() 
m2 |> plot_fa_()
 
}
\keyword{internal}
